==> 07b_repport_classify.py <==
# Script for generating reports on drought classification
# usage : python 06b_repport_classify.py <dataset>
import os
import sys

import pandas as pd
from alive_progress import alive_bar
import ast
import json
from sklearn.metrics import confusion_matrix, ConfusionMatrixDisplay
from sklearn.metrics import precision_recall_fscore_support,accuracy_score

import jinja2
import io
import base64

from matplotlib import pyplot as plt
import matplotlib.patches as mpatches
import numpy as np

tipo_agro="agrocultura"
tipo_ganaderia="ganaderia"
tipo_hidrologia="hidrologia"
tipo_energia="energia"
tipos = [tipo_agro, tipo_ganaderia, tipo_hidrologia, tipo_energia]

ciena_tvi={"qwen25.72b.cot":{"accuracy":0.742,"precision":0.737,"recall":0.832,"f1_score":0.782,"time": 30.07},
            "qwen25.7b":{"accuracy":0.686,"precision":0.791,"recall":0.544,"f1_score":0.645,"time": 2.54},
            "qwen25.3b":{"accuracy":0.634,"precision":0.769,"recall":0.320,"f1_score":0.452,"time": 1.811}
            }

def test_file_name_to_model(file_name):
    
    # extremadura_1962_9_1_page_1 -> extremadura_1962_09_01_page_0001
    parts = file_name.split('_')
    parts[2] = parts[2].zfill(2)
    parts[3] = parts[3].zfill(2)
    parts[5] = parts[5].zfill(4)
    return '_'.join(parts)

def extract_data(row, field):
    val = row.get(field, 0)
    # Default is False
    if val in [0, '0', False, 'False', 'false', None]: return False
    if val in [1, '1', True, 'True', 'true']: return True
    
    print (f"Unknown value for {field}: {val}")
    exit()

def extract_row(pred_row, real_ds):
            file_name = pred_row['article_filename'].split('/')[-1]
            file_name = file_name.replace('.json', '')
            real_row = real_ds[real_ds['file_name'] == file_name]

            if real_row.empty:
                # try with model file name
                model_file_name = test_file_name_to_model(file_name)
                real_row = real_ds[real_ds['file_name'] == model_file_name]

            if real_row.empty:
                print(f"File not found in real data: {file_name}")
                return None
            
            real_sequia = real_row.iloc[0]['has_sequia']
            if not real_sequia:
                return None  # Skip non-drought articles

            row_sequia = pred_row['article_extracted_data']
            if row_sequia.startswith("{'"):
                row_sequia = ast.literal_eval(row_sequia)
            else:
                row_sequia = json.loads(row_sequia)

            #agrocultura,ganaderia,hidrologia,energia
            real_agrocultura = extract_data(real_row.iloc[0], tipo_agro)
            real_ganaderia = extract_data(real_row.iloc[0], tipo_ganaderia)
            real_hidrologia = extract_data(real_row.iloc[0], tipo_hidrologia)
            real_energia = extract_data(real_row.iloc[0], tipo_energia)
            
            pred_agrocultura = extract_data(row_sequia, "agriculture")
            pred_ganaderia = extract_data(row_sequia, "livestock")
            pred_hidrologia = extract_data(row_sequia, "hydrological_resources")
            pred_energia = extract_data(row_sequia, "energy")

            #print (row_sequia)
            return {
                'file_name': file_name,
                'real_'+tipo_agro: real_agrocultura,
                'real_'+tipo_ganaderia: real_ganaderia,
                'real_'+tipo_hidrologia: real_hidrologia,
                'real_'+tipo_energia: real_energia,
                'pred_'+tipo_agro: pred_agrocultura,
                'pred_'+tipo_ganaderia: pred_ganaderia,
                'pred_'+tipo_hidrologia: pred_hidrologia,
                'pred_'+tipo_energia: pred_energia,         
            }


def build_ds_compare(real_ds, pred_ds): 
    with alive_bar(len(pred_ds), title=f'Building comparison dataset') as bar:
        merged = []
        for _, pred_row in pred_ds.iterrows():
            row_data= extract_row(pred_row, real_ds)
            if row_data is not None:
                merged.append(row_data)
            bar()
    return pd.DataFrame(merged)

def generate_reports(real_ds, dataset, test_name):
    if not os.path.exists(f"results/{dataset}/classify/{test_name}/summary.csv"):
        return None
    
    pred_ds = pd.read_csv(f"results/{dataset}/classify/{test_name}/summary.csv")
    df = build_ds_compare(real_ds, pred_ds)

    print("---- Report ----")
    print (f"Report for test: {test_name}")
    print(f"Number of Agro detected: {df['pred_'+tipo_agro].sum()} out of {len(df)} ")
    print(f"Number of Real Agro: {df['real_'+tipo_agro].sum()} out of {len(df)} ") 
    #print (df.head())
    
    # Make confusion matrix
    info = {}
    for tipo in tipos:
        print(f"Confusion Matrix for {tipo}:")
        print (f"Predicted positives: {df['pred_'+tipo].sum()}, Real positives: {df['real_'+tipo].sum()}")
        cm = confusion_matrix(df['real_'+tipo], df['pred_'+tipo],labels=[True, False])
        print(cm)
        fig, ax = plt.subplots(figsize=(4, 2.5),layout="constrained")
        disp = ConfusionMatrixDisplay(confusion_matrix=cm, display_labels=[tipo, f'No {tipo}'])
        disp.plot(ax=ax, text_kw={'fontsize':14}, cmap=plt.cm.Blues)
        ax.set_ylabel("Real", fontsize=10)
        ax.set_xlabel("Predicción", fontsize=10)
        #save to memory
        buf = io.BytesIO()
        disp.figure_.savefig(buf, format='png')
        buf.seek(0)
        b64_img = base64.b64encode(buf.read()).decode('utf-8')
        plt.close(disp.figure_)

        accuracy = accuracy_score(df['real_'+tipo], df['pred_'+tipo])
        print(f"Accuracy: {accuracy:.2f}")
        score = precision_recall_fscore_support(df['real_'+tipo], df['pred_'+tipo],labels=[True])
        print("Precision, Recall, F1-Score:")
        print(score)     
        
        
        info[tipo] = {
            "name": tipo,
            "accuracy": accuracy,
            "precision": score[0][0],
            "recall": score[1][0],
            "f1_score": score[2][0],
            "cm_image": b64_img
        }
    
    # Display stats (accuracy, precision, recall, f1-score )º
    accuracy =0
    precision =0
    recall =0
    f1_score =0
    for tipo in tipos:
        accuracy += info[tipo]["accuracy"]
        precision += info[tipo]["precision"]
        recall += info[tipo]["recall"]
        f1_score += info[tipo]["f1_score"]
    
    accuracy /= len(tipos)
    precision /= len(tipos)
    recall /= len(tipos)
    f1_score /= len(tipos)

    times_js=json.loads(open(f"results/{dataset}/classify/{test_name}/execution_times.json").read())
    full_time=times_js.get("total",None)
    avg_time=full_time/len(df) if full_time is not None else None
    parrsing_errors_js= json.loads(open(f"results/{dataset}/classify/{test_name}/parsing_errors.json").read())
    parsing_errors=parrsing_errors_js.get("total",0)

    print(f"Total Execution Time: {full_time:.2f} seconds" if full_time is not None else "Total Execution Time: N/A")
    print(f"Average Time per Article: {avg_time:.2f} seconds" if avg_time is not None else "Average Time per Article: N/A")
    print(f"Total Parsing Errors: {parsing_errors}")

    print("---- End Report ----")
    return {
        'test_name': test_name,
        'accuracy': accuracy,
        'precision': precision,
        'recall': recall,
        'f1_score': f1_score,
        'total': len(df),
        'tipos_info': info,
        "full_time": full_time,
        "avg_time": avg_time,
        "parsing_errors": parsing_errors
    }

def build_table(data, key):
    table = {}
    for row in data:
        # model = row['test_name'].split('-')[0]
        # mode = row['test_name'].replace(f'{model}-','')
        # if model not in table:
        #     table[model] = {'model': model}
        # table[model][mode] = row[key]
        name = row['test_name']
        table[name] = {"model":name,"global": row[key]}
        for tipo in tipos:
            if key not in row['tipos_info'][tipo]:
                continue
            table[name][tipo] = row['tipos_info'][tipo][key]

    return pd.DataFrame.from_dict(table, orient='index')

def lighten_color(color, amount=0.5):
    """
    Lightens the given color by multiplying (1-luminosity) by the given amount.
    Input can be matplotlib color string, hex string, or RGB tuple.

    Examples:
    >> lighten_color('g', 0.3)
    >> lighten_color('#F034A3', 0.6)
    >> lighten_color((.3,.55,.1), 0.5)
    """
    import matplotlib.colors as mc
    import colorsys
    try:
        c = mc.cnames[color]
    except:
        c = color
    c = colorsys.rgb_to_hls(*mc.to_rgb(c))
    return colorsys.hls_to_rgb(c[0], 1 - amount * (1 - c[1]), c[2])


base_colors = {
        'qwen25.3b': '#E91E63',
        'qwen25.7b': '#FF9800',
        'deepseek.8b': "#3C4965",
        'qwen3.8b': "#7A980C",
        'qwen3.32b.cot': '#1E5AA8',
        'qwen25.72b.cot': '#2ECC71'
    }

colors = {
        
    "qwen25.3b-no-summary": base_colors['qwen25.3b'],
    "qwen25.3b-summary": lighten_color(base_colors['qwen25.3b'],0.4),
    "qwen25.3b-summary-expert": lighten_color(base_colors['qwen25.3b'],0.2),

    "qwen25.7b-no-summary": base_colors["qwen25.7b"],
    "qwen25.7b-summary": lighten_color(base_colors["qwen25.7b"],0.4),
    "qwen25.7b-summary-expert": lighten_color(base_colors["qwen25.7b"],0.2),

    "deepseek.8b-no-summary": base_colors["deepseek.8b"],
    "deepseek.8b-summary": lighten_color(base_colors["deepseek.8b"],0.4),
    "deepseek.8b-summary-expert": lighten_color(base_colors["deepseek.8b"],0.2),

    "qwen3.8b-no-summary": base_colors["qwen3.8b"],
    "qwen3.8b-summary": lighten_color(base_colors["qwen3.8b"],0.4),
    "qwen3.8b-summary-expert": lighten_color(base_colors["qwen3.8b"],0.2),

    "qwen3.32b.cot-no-summary": base_colors["qwen3.32b.cot"],
    "qwen3.32b.cot-summary": lighten_color(base_colors["qwen3.32b.cot"],0.4),
    "qwen3.32b.cot-summary-expert": lighten_color(base_colors["qwen3.32b.cot"],0.2),

    "qwen25.72b.cot-no-summary": base_colors["qwen25.72b.cot"],
    "qwen25.72b.cot-summary": lighten_color(base_colors["qwen25.72b.cot"],0.4),
    "qwen25.72b.cot-summary-expert": lighten_color(base_colors["qwen25.72b.cot"],0.2),
}   

def build_bar_chart(df, name):
    groups = df.columns[1:].to_list()
    models = df["model"].to_list()

    groups = [group for group in groups if group != "global"]
    print ("Building bar chart...")

    group_names={"agrocultura":"Agricultura","ganaderia":"Ganadería","hidrologia":"Hidrología","energia":"Energía"}

    # setup 4 subplots
    fig, axes = plt.subplots(2, 2, figsize=(12, 8), layout='constrained', sharex=True, sharey=True)
    axes = axes.flatten()
    for i, grupo in enumerate(groups):
        ax = axes[i]

        if name == "F1-Score":
            ax.axhline(y=ciena_tvi['qwen25.3b']["f1_score"], color='gray', linestyle='dotted',alpha=0.7)
            ax.axhline(y=ciena_tvi['qwen25.7b']["f1_score"], color='gray', linestyle='dashed',alpha=0.7)
            ax.axhline(y=ciena_tvi['qwen25.72b.cot']["f1_score"], color='gray', linestyle='dotted',alpha=0.7)

        for model in models:
            value = df[df["model"]==model].iloc[0][grupo]
            if pd.isna(value):
                continue
            rects = ax.bar(model, value, color=colors.get(f"{model}", None))
            #ax.bar_label(rects, padding=3)
        ax.set_title(group_names.get(grupo, grupo.capitalize()))
        ax.set_ylim(0, 1)
        ax.set_ylabel(name)
        plt.setp(ax.get_xticklabels(), rotation=45, ha='right', rotation_mode='anchor')
        #ax.xticks(rotation=45, ha='right', rotation_mode='anchor')
        #plt.xticks(rotation=45, ha='right', rotation_mode='anchor')
        
    handles = [mpatches.Patch(color=colors[model], label=model) for model in models]
    fig.legend(handles=handles, loc='outside right center', ncol=1)
    #plt.xticks(rotation=45, ha='right', rotation_mode='anchor')
    plt.suptitle(f'{name} por Tipo de Sequía', fontsize=16)
    
    buf = io.BytesIO()
    fig.savefig(buf, format='png')
    buf.seek(0)
    b64_img = base64.b64encode(buf.read()).decode('utf-8')
    plt.close(fig)
    return b64_img


def cross_table(df, key, include_ciena=True):
    table = {}
    models =[]
    for _, row in df.iterrows():
        model = row['model'].split('-')[0]
        mode = row['model'].replace(f'{model}-','')
        if model not in table:
            table[model] = {'model': model}
        table[model][mode] = row[key]
        if model not in models:
            models.append(model)
    if include_ciena:
        for model in models:
            if model not in ciena_tvi:
                continue
            table[model]["ciena"] = ciena_tvi[model]["f1_score"]
    
    return pd.DataFrame.from_dict(table, orient='index')



def sort_tests(tests):
    def sort_key(test_name):
        parts = test_name.split('-')
        model = parts[0]
        mode = '-'.join(parts[1:])
        model_order = {'qwen25.72b.cot': 0, 'qwen25.7b': 1, 'qwen25.3b': 2, 'deepseek.8b': 3}
        model_order= {"qwen25.3b":0,"qwen25.7b":1,"deepseek.8b":2,"qwen3.8b":3,"qwen3.32b.cot":4,"qwen25.72b.cot":5}        
        mode_order = {'no-summary': 0, 'summary': 1, 'summary-expert': 2}

        return (model_order.get(model, 99), mode_order.get(mode, 99))
    
    return sorted(tests, key=sort_key)


def plot_f1_scores(df_f1:pd.DataFrame, times:pd.DataFrame):
    # Plot F1 scores with execution times
    # color for model

    markers = {
        'no-summary': 'o',
        'summary': 's',
        'summary-expert': 'D'
    }
    fig, ax1 = plt.subplots(figsize=(10, 5))

    # ax1.axhline(y=ciena_tvi['qwen25.3b']["f1_score"], color='gray', linestyle='dotted',alpha=0.7)
    # ax1.axhline(y=ciena_tvi['qwen25.7b']["f1_score"], color='gray', linestyle='dashed',alpha=0.7)
    # ax1.axhline(y=ciena_tvi['qwen25.72b.cot']["f1_score"], color='gray', linestyle='dotted',alpha=0.7)
    
    ax1.axvline(x=ciena_tvi['qwen25.3b']["time"], color='gray', linestyle='dotted',alpha=0.7)
    ax1.axvline(x=ciena_tvi['qwen25.7b']["time"], color='gray', linestyle='dashed',alpha=0.7)
    ax1.axvline(x=ciena_tvi['qwen25.72b.cot']["time"], color='gray', linestyle='dotted',alpha=0.7)


    for model in df_f1['model']:
        model_data = df_f1[df_f1['model'] == model]
        for mode in ['no-summary', 'summary', 'summary-expert']:
            if mode in model_data.columns:
                ax1.scatter(times.loc[model, mode],
                model_data[mode],
                label=f"{model} - {mode}",
                color=base_colors.get(model, 'gray'),
                marker=markers.get(mode, 'o'))
    ax1.set_ylabel('F1 Score')
    ax1.set_xlabel('Tiempo por artículo (s)')
    # Leyenda personalizada
    # colores únicos para modelos
    handles = [plt.Line2D([0], [0], marker='None', color='w', label='Modelo')] 
    for model in df_f1['model']:
        handles.append(plt.Line2D([0], [0], marker='o', color='w', label=model, markerfacecolor=base_colors.get(model, 'gray'), markersize=10))
    # Agregar separación entre modelos y modos
    handles.append(plt.Line2D([0], [0], marker='None', color='w', label=''))
    handles.append(plt.Line2D([0], [0], marker='None', color='w', label='Resumen'))
    # marcadores únicos para modos 
    for mode in ['no-summary', 'summary', 'summary-expert']:
        handles.append(plt.Line2D([0], [0], marker=markers.get(mode, 'o'), color='w', label=mode, markerfacecolor='black', markersize=10))
    ax1.legend(handles=handles, loc='right', bbox_to_anchor=(1.25, 0.5))
    plt.title('F1 Score en Clasificación de Sequías')
    plt.tight_layout()
    # plt.show()
    buf = io.BytesIO()
    plt.savefig(buf, format='png')  
    buf.seek(0)
    b64_img = base64.b64encode(buf.read()).decode('utf-8')
    plt.close(fig)
    return b64_img

def calc_diff_summary(df_f1):
    print("Calculating difference summary...")

    def calc_diff(row):
        name = row.name
        base = row.get('no-summary')
        summary = row.get('summary')
        summary_expert = row.get('summary-expert', None)
        ret = {'summary': summary - base}
        if summary_expert is not None:
            ret['summary_expert'] = summary_expert - base

        return pd.Series(ret, name=name)
    return df_f1.apply(calc_diff, axis=1)

def plot_diff_summary(diff_summary):
    fig, ax = plt.subplots(figsize=(6, 4))
    diff_summary.plot(kind='bar', ax=ax)
    ax.set_title('Cambios F1 Score tras aplicar resúmenes')
    ax.set_ylabel('Diferencia F1-Score')
    ax.axhline(0, color='gray', linestyle='--')
    plt.xticks(rotation=45, ha='right', rotation_mode='anchor')
    plt.tight_layout()
    #plt.show()
    buf = io.BytesIO()
    plt.savefig(buf, format='png')
    buf.seek(0)
    b64_img = base64.b64encode(buf.read()).decode('utf-8')
    plt.close(fig)
    return b64_img


def main():
    if len(sys.argv) != 2:
        print("Usage: python 06b_repport_classify.py <dataset>")
        sys.exit(1)

    dataset = sys.argv[1]
    tests = os.listdir(f"results/{dataset}/classify/")
    tests = [test for test in tests if os.path.isdir(f"results/{dataset}/classify/{test}")]
    real_ds = pd.read_csv(f"data/datasets/{dataset}_ds/classify/{dataset}_ds.csv")

    for tipo in tipos:
        print (f"Total real {tipo}: {real_ds[tipo].sum()} out of {len(real_ds)}")

    data = []

    tests = sort_tests(tests)

    for test in tests:
        print(f"Generating report for test: {test}")
        tdata=generate_reports(real_ds, dataset, test)
        if tdata is not None:
            data.append(tdata)
    
    df_acc=build_table(data, 'accuracy')
    df_f1=build_table(data, 'f1_score')
    avg_time = cross_table(build_table(data, 'avg_time'), 'global', include_ciena=False)
    errors = cross_table(build_table(data, 'parsing_errors'), 'global', include_ciena=False)
    df_f1_global = cross_table(df_f1, 'global')
    print("=== Accuracy Table ===")
    print(df_acc)
    print("=== F1-Score Table ===")
    print(df_f1)
    print("=== F1-Score Global Table ===")
    print(df_f1_global)

    img_f1_global = plot_f1_scores(df_f1_global, avg_time)

    print("=== Average Time per Article ===")
    print(avg_time)
    print("=== Parsing Errors ===")
    print(errors)

    diff_summary  = calc_diff_summary(df_f1_global)
    print("=== Difference Summary ===")
    print(diff_summary)
    diff_summary_img = plot_diff_summary(diff_summary)

    renderer = jinja2.Environment(
        loader=jinja2.FileSystemLoader(searchpath="./data/templates/")
    )
    template = renderer.get_template("classify.html.jinja")
    output = template.render(
        dataset=dataset,
        accuracy_table=df_acc.to_html(index=False, float_format="{:.2f}".format),
        f1_score_table=df_f1.to_html(index=False, float_format="{:.2f}".format),
        f1_global=df_f1_global.to_html(index=False, float_format="{:.2f}".format),
        average_time_table=avg_time.to_html(index=False, float_format="{:.2f}".format),
        parsing_errors_table=errors.to_html(index=False),
        results=data,
        accuracy_chart=build_bar_chart(df_acc, "Accuracy"),
        f1_score_chart=build_bar_chart(df_f1, "F1-Score"),
        img_f1_global=img_f1_global,
        diff_summary_image=diff_summary_img
    )

    with open(f"results/{dataset}/classify/report-classify.html", "w", encoding='utf-8') as f:
        f.write(output)

    
    for _, row in df_f1_global.iterrows():
        print(f"\"{row['model']}-no-summary\": {row['no-summary']:.3f},")
        print(f"\"{row['model']}-summary\": {row['summary']:.3f},")
        if 'summary-expert' in row:
            print(f"\"{row['model']}-summary-expert\": {row['summary-expert']:.3f},")


if __name__ == "__main__":
    main()
